+++ app/controllers/skaters_controller.rb
@@ -1,79 +1,58 @@
-################
-class SkatersListDecorator < ListDecorator
-  def name
-    h.link_to_skater(model)
+class SkaterCompetitionResultSummary
+  extend Forwardable
+  def_delegators :@result_summary, :[]
+  
+  def initialize(skater, isu_championships_only: false)
+    @skater = skater
+    @category_results = skater.category_results.isu_championships_only_if(isu_championships_only)
+    @isu_champions_only = isu_championships_only
+=begin
+    collection = skater.category_results.isu_championships_only_if(isu_championships_only)
+    tmp_pcs = Hash.new { |h,k| h[k] = []}
+    skater.components.isu_championships_only_if(isu_championships_only).each {|c| tmp_pcs[c.number] << c.value}
+    max_pcs = {}
+    (1..5).each do |i|
+      max_pcs[i] = tmp_pcs[i].max
+    end
+    binding.pry
+    @result_summary = {
+      highest_score: collection.pluck(:points).compact.max,
+      competitions_participated: collection.count,
+      gold_won: collection.where(ranking: 1).count,
+      highest_ranking: collection.pluck(:ranking).compact.reject {|d| d == 0}.min,
+      most_valuable_element: skater.elements.isu_championships_only_if(isu_championships_only).order(:value).last,
+      most_valuable_components: max_pcs || {},
+    }
+=end
   end
-  def isu_number
-    h.link_to_isu_bio(model.isu_number)
+  def highest_score
+    @category_results.pluck(:points).compact.max
   end
-end
+  def competitions_participated
+    @category_results.count
+  end
+  def gold_won
+    @category_results.where(ranking: 1).count
+  end
+  def highest_ranking
+    @category_results.pluck(:ranking).compact.reject {|d| d == 0}.min
+  end
 
-class SkaterCompetitionsListDecorator < ListDecorator
-  include ApplicationHelper
-  class << self
-    def headers
-      {
-        short_ranking: "SP#",
-        short_tss: "SP-TSS",
-        short_tes: "SP-TES",
-        short_pcs: "SP-PCS",
-        short_deductions: "SP-ded",
-        free_ranking: "FS#",
-        free_tss: "FS-TSS",
-        free_tes: "FS-TES",
-        free_pcs: "FS-PCS",
-        free_deductions: "FS-ded",
-      }
+  def most_valuable_element
+    if (elem = @skater.elements.isu_championships_only_if(@isu_championships_only).order(:value).last)
+      "%s %s%s (%.2f=%.2f+%.2f)" % [ elem.element, elem.credit, elem.info, elem.value, elem.base_value, elem.goe]
+    else
+      "-"
     end
   end
-  def competition_name
-    h.link_to_competition(nil, model.competition)
+  def most_valuable_components
+    @skater.components.isu_championships_only_if(@isu_championships_only).group(:number).maximum(:value).values.join('/')
   end
-  def date
-    model.competition.start_date
-  end
-  def ranking
-    h.link_to_competition(as_ranking(model.ranking), model.competition, category: model.category)
-  end
-  def points
-    h.link_to_competition(as_score(model.points), model.competition, category: model.category)
-  end
-  ## short
-  def short_ranking
-    h.link_to_score(as_ranking(model.short_ranking), model.scores.first)
-  end
-  def short_tss
-    (s = model.scores.first) ? as_score(s.tss) : "-"
-  end
-  def short_tes
-    (s = model.scores.first) ? as_score(s.tes) : "-"
-  end
-  def short_pcs
-    (s = model.scores.first) ? as_score(s.pcs) : "-"    
-  end
-  def short_deductions
-    (s = model.scores.first) ? as_score(s.deductions) : "-"    
-  end
-
-  ## free
-  def free_ranking
-    h.link_to_score(as_ranking(model.free_ranking), model.scores.first)
-  end
-  def free_tss
-    (s = model.scores.second) ? as_score(s.tss) : "-"
-  end
-  def free_tes
-    (s = model.scores.second) ? as_score(s.tes) : "-"
-  end
-  def free_pcs
-    (s = model.scores.second) ? as_score(s.pcs) : "-"
-  end
-  def free_deductions
-    (s = model.scores.second) ? as_score(s.deductions) : "-"
-  end
-
 end
-################################################################
+class SkaterCompetitionResultSummaryDecorator < EntryDecorator
+end
+
+################
 class SkatersController < ApplicationController
   ## index
   def filters
@@ -102,37 +81,32 @@ class SkatersController < ApplicationController
   def show_skater(skater)
     raise ActiveRecord::RecordNotFound.new("no such skater") if skater.nil?
 
+    ################
+    ## competition results
     collection = skater.category_results.with_competition.recent.includes(:competition, :scores).isu_championships_only_if(params[:isu_championships_only])
 
-    category_results = SkaterCompetitionsListDecorator.decorate_collection(collection)
+    competition_results = SkaterCompetitionDecorator.decorate_collection(collection)
 
+    ################
     ## result summary
-    tmp_pcs = Hash.new { |h,k| h[k] = []}
-    skater.components.all.each {|c| tmp_pcs[c.number] << c.value}
-    max_pcs = {}
-    (1..5).each do |i|
-      max_pcs[i] = tmp_pcs[i].max
-    end
-    result_summary = {
-      highest_score: collection.pluck(:points).compact.max,
-      competitions_participated: collection.count,
-      gold_won: collection.where(ranking: 1).count,
-      highest_ranking: collection.pluck(:ranking).compact.reject {|d| d == 0}.min,
-      most_valuable_element: skater.elements.order(:value).last,
-      most_valuable_components: max_pcs || {},
-                                        
-    }
-    #collection = skater.category_results.includes(:competition)
-    #category_results = SkaterCompetitionsListDecorator.decorate_collection(collection.includes(:scores).order("scores.date desc"))
-
+    result_summary = SkaterCompetitionResultSummary.new(skater, isu_championships_only: params[:isu_championships_only])
+                                                        
+    ################
+    ## score graph
     score_graph = ScoreGraph.new
     [:short, :free].each do |segment_type|
       score_graph.plot(skater, skater.scores.send(segment_type).recent.isu_championships_only_if(params[:isu_championships_only]), segment_type)
     end
 
+    ################
+    ## render
     respond_to do |format|
-      format.html { render action: :show, locals: { skater: skater, category_results: category_results, result_summary: result_summary }}
-      format.json { render json: {skater_info: skater, competition_results: skater.category_results} }
+      format.html {
+        render action: :show, locals: { skater: skater, competition_results: competition_results, result_summary: result_summary }
+      }
+      format.json {
+        render json: {skater_info: skater, competition_results: skater.category_results}
+      }
     end
   end
 end