+++ app/decorators/skater_decorator.rb
@@ -1,0 +1,9 @@
+class SkaterDecorator < EntryDecorator
+  def name
+    h.link_to_skater(model)
+  end
+  def isu_number
+    h.link_to_isu_bio(model.isu_number)
+  end
+end
+