+++ app/models/component.rb
@@ -1,4 +1,6 @@
 class Component < ApplicationRecord
+  include IsuChampionshipsOnly
+  
   belongs_to :score
 
   scope :recent, ->{ with_score.order("scores.date desc") }