+++ app/models/element.rb
@@ -1,6 +1,9 @@
 class Element < ApplicationRecord
+  include IsuChampionshipsOnly
   belongs_to :score
 
   scope :recent, ->{ with_score.order("scores.date desc") }
   scope :with_competition, ->{ joins(score: [:competition]) }
+
+  
 end